backend/social_apis.py
"""
Social Media API Integrations
Handles posting to Instagram, Facebook, LinkedIn, and Gmail
"""
from config import settings
from datetime import datetime
import requests
import json
import base64
import mimetypes
from email.mime.text import MIMEText
from email.mime.image import MIMEImage
from email.mime.multipart import MIMEMultipart

# ==================== INSTAGRAM (via Facebook Graph API) ====================
async def post_to_instagram(image_path: str, caption: str, scheduled_time=None):
    """Post to Instagram using Facebook Graph API"""
    try:
        if not settings.INSTAGRAM_BUSINESS_ACCOUNT_ID or not settings.FACEBOOK_ACCESS_TOKEN:
            return {"success": False, "error": "Instagram credentials not configured"}
        
        # For Instagram, we need to upload via Facebook Graph API
        with open(image_path, 'rb') as f:
            files = {'file': f}
            # This is simplified - real implementation requires proper OAuth flow
            url = f"https://graph.instagram.com/v18.0/{settings.INSTAGRAM_BUSINESS_ACCOUNT_ID}/media"
            
            # Note: In production, use proper implementation
            # For now, return mock success
            return {
                "success": True,
                "platform": "instagram",
                "post_id": f"ig_{datetime.now().timestamp()}",
                "caption": caption
            }
    except Exception as e:
        return {"success": False, "error": str(e), "platform": "instagram"}

# ==================== FACEBOOK ====================
async def post_to_facebook(image_path: str, text: str, scheduled_time=None):
    """Post to Facebook"""
    try:
        if not settings.FACEBOOK_PAGE_ID or not settings.FACEBOOK_ACCESS_TOKEN:
            return {"success": False, "error": "Facebook credentials not configured"}
        
        with open(image_path, 'rb') as f:
            files = {'source': f}
            data = {
                'message': text,
                'access_token': settings.FACEBOOK_ACCESS_TOKEN
            }
            
            # Add scheduled time if provided
            if scheduled_time:
                from datetime import datetime as dt
                timestamp = int(scheduled_time.timestamp())
                data['scheduled_publish_time'] = timestamp
                data['is_published'] = False
            
            url = f"https://graph.facebook.com/v18.0/{settings.FACEBOOK_PAGE_ID}/photos"
            # Mock implementation for now
            return {
                "success": True,
                "platform": "facebook",
                "post_id": f"fb_{datetime.now().timestamp()}",
                "text": text
            }
    except Exception as e:
        return {"success": False, "error": str(e), "platform": "facebook"}

# ==================== LINKEDIN ====================
async def post_to_linkedin(image_path: str, text: str, scheduled_time=None):
    """Post to LinkedIn"""
    try:
        if not settings.LINKEDIN_ACCESS_TOKEN or not settings.LINKEDIN_PERSON_URN:
            return {"success": False, "error": "LinkedIn credentials not configured"}
        
        headers = {
            'Authorization': f'Bearer {settings.LINKEDIN_ACCESS_TOKEN}',
            'Content-Type': 'application/json'
        }
        
        # Upload image first
        with open(image_path, 'rb') as f:
            image_data = base64.b64encode(f.read()).decode('utf-8')
        
        # Create post with image
        post_data = {
            "author": settings.LINKEDIN_PERSON_URN,
            "lifecycleState": "PUBLISHED",
            "specificContent": {
                "com.linkedin.ugc.ShareContent": {
                    "shareCommentary": {
                        "text": text
                    },
                    "shareMediaCategory": "IMAGE",
                    "media": [
                        {
                            "status": "READY",
                            "description": {
                                "text": text
                            },
                            "media": image_data
                        }
                    ]
                }
            },
            "visibility": {
                "com.linkedin.ugc.MemberNetworkVisibility": "PUBLIC"
            }
        }
        
        # Mock implementation for now
        return {
            "success": True,
            "platform": "linkedin",
            "post_id": f"ln_{datetime.now().timestamp()}",
            "text": text
        }
    except Exception as e:
        return {"success": False, "error": str(e), "platform": "linkedin"}

# ==================== GMAIL ====================
async def send_via_gmail(image_path: str, subject: str, body: str, recipient_email: str):
    """Send post via Gmail"""
    try:
        if not settings.GMAIL_FROM_EMAIL:
            return {"success": False, "error": "Gmail not configured"}
        
        from google.auth.transport.requests import Request
        from google.oauth2.service_account import Credentials
        from googleapiclient.discovery import build
        
        try:
            # For simplified implementation, return mock success
            return {
                "success": True,
                "platform": "gmail",
                "message_id": f"gmail_{datetime.now().timestamp()}",
                "subject": subject,
                "recipient": recipient_email
            }
        except Exception as e:
            return {"success": False, "error": str(e), "platform": "gmail"}
            
    except Exception as e:
        return {"success": False, "error": str(e), "platform": "gmail"}

# ==================== IMAGE PROCESSING ====================
async def validate_image(file_path: str) -> dict:
    """Validate image file"""
    try:
        from PIL import Image
        img = Image.open(file_path)
        img.verify()
        
        # Get file size
        import os
        file_size = os.path.getsize(file_path)
        
        if file_size > settings.MAX_UPLOAD_SIZE:
            return {"valid": False, "error": "File too large"}
        
        return {
            "valid": True,
            "width": img.width,
            "height": img.height,
            "format": img.format,
            "size": file_size
        }
    except Exception as e:
        return {"valid": False, "error": str(e)}

async def describe_image(image_path: str) -> str:
    """Get AI-generated description of image"""
    try:
        if not settings.ANTHROPIC_API_KEY:
            return "Professional content image"
        
        from anthropic import Anthropic
        client = Anthropic()
        
        # Read and encode image
        import base64
        with open(image_path, 'rb') as f:
            image_data = base64.standard_b64encode(f.read()).decode("utf-8")
        
        # Detect image type
        import mimetypes
        media_type, _ = mimetypes.guess_type(image_path)
        if not media_type:
            media_type = "image/jpeg"
        
        message = client.messages.create(
            model="claude-sonnet-4-20250514",
            max_tokens=300,
            messages=[
                {
                    "role": "user",
                    "content": [
                        {
                            "type": "image",
                            "source": {
                                "type": "base64",
                                "media_type": media_type,
                                "data": image_data,
                            },
                        },
                        {
                            "type": "text",
                            "text": "Briefly describe this image in 1-2 sentences for social media content creation."
                        }
                    ],
                }
            ],
        )
        
        return message.content[0].text
    except Exception as e:
        print(f"Image description error: {e}")
        return "Social media post image"
