backend/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
anthropic
python-dotenv
pydantic
pillow
aiofiles
python-multipart
facebook-sdk
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
instagrapi
linkedin-api
schedule
httpx
sqlalchemy
pydantic-settings
